+++ rust/azure_iot_operations_connector/src/base_connector.rs
@@ -12,7 +12,7 @@ use azure_iot_operations_protocol::application::ApplicationContext;
 use azure_iot_operations_services::{azure_device_registry, schema_registry, state_store};
 use managed_azure_device_registry::DeviceEndpointClientCreationObservation;
 
-use crate::filemount::connector_artifacts::ConnectorArtifacts;
+use crate::deployment_artifacts::connector::ConnectorArtifacts;
 
 pub mod adr_discovery;
 pub mod managed_azure_device_registry;
