+++ rust/azure_iot_operations_connector/src/deployment_artifacts/azure_device_registry.rs
