--- rust/azure_iot_operations_connector/src/filemount.rs
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-//! Utilities for extracting data from filemounts in an Akri deployment
-
-pub mod azure_device_registry;
-pub mod connector_artifacts;
-pub mod observability_artifacts;
-
-// TODO: Add common artifact structs and helpers here once implementation is unified
