+++ rust/azure_iot_operations_connector/src/lib.rs
@@ -13,8 +13,8 @@ use azure_iot_operations_services::{
 
 pub mod base_connector;
 pub mod data_processor;
+pub mod deployment_artifacts;
 pub mod destination_endpoint;
-pub mod filemount;
 pub mod source_endpoint;
 
 #[macro_use]
